Moved file parsing functions into a new file
